implement defense feedback
